reflection-agent/pyproject.toml
[project]
name = "reflection-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Sandeep Kumar Gupta",email = "sangupta019@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-openai (>=1.1.9,<2.0.0)",
    "langgraph (>=1.0.8,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
